Moved sea creatures to own dir
